scripts(pkglist): adjust some seq and add some packs scripts(misc): improve buildall.sh graphviz(misc): improve xfce and add fcitx

+++ scripts/buildall.sh
@@ -4,9 +4,9 @@ GREEN='\e[1;32m'
 YELLOW='\e[1;33m'
 END='\e[0m'
 PASSWORD='password'
-RPMPATH='~/rpmbuild/RPMS'
-SRCPATH='~/sources/src-openeuler/pkgs'
-PKGLIST='./pkglist'
+RPMPATH='/rpmbuild/RPMS'
+SRCPATH='/sources/src-openeuler/pkgs'
+PKGLIST='/pkglist'
 
 checkStatus() {
     if [ $1 == 0 ]
@@ -28,36 +28,59 @@ add_repo() {
 
 install_tools
 add_repo
+succ=0
+fail=0
+skip=0
 while read pkgname;
 do
     cd ${SRCPATH}/${pkgname}
-    echo -e "${YELLOW}======= Stage 1: Copy Build Files =======${END}"
+    count1=`find ${RPMPATH}/aarch64/ -type f | wc -l`
+    count1=$[count1 + `find ${RPMPATH}/noarch/ -type f | wc -l`]
+    if [ -f ${SRCPATH}/${pkgname}/success.token ]; then
+        echo -e "${YELLOW}======= Stage 1: Copy Build Files =======${END}"
 
-    status=0
-    find ./ -type f -not -iname "*.spec" -not -iname "*.md" -not -iname "*.yaml" -not -path '*/\.*' -exec cp {} ~/rpmbuild/SOURCES/ \;
-    cp *.spec ~/rpmbuild/SPECS/
-    tmp=$?
-    status=$[tmp + status];
-    rm ~/rpmbuild/SOURCES/*.spec ~/rpmbuild/SOURCES/*.md
-    tmp=$?
-    status=$[tmp + status];
-    checkStatus $status
+        status=0
+        find ./ -type f -not -iname "*.spec" -not -iname "*.md" -not -iname "*.yaml" -not -path '*/\.*' -exec cp {} ~/rpmbuild/SOURCES/ \;
+        cp *.spec ~/rpmbuild/SPECS/
+        tmp=$?
+        status=$[tmp + status];
+        rm ~/rpmbuild/SOURCES/*.spec ~/rpmbuild/SOURCES/*.md
+        tmp=$?
+        status=$[tmp + status];
+        checkStatus $status
 
-    echo -e "${YELLOW}========== Stage 2: Build Deps ==========${END}"
+        echo -e "${YELLOW}========== Stage 2: Build Deps ==========${END}"
 
-    status=0
-    echo ${PASSWORD} | sudo -S dnf -y builddep *.spec --refresh
-    status=$[tmp + status];
-    checkStatus $status
+        status=0
+        echo ${PASSWORD} | sudo -S dnf -y builddep *.spec --refresh
+        status=$[tmp + status];
+        checkStatus $status
 
-    echo -e "${YELLOW}======== Stage 3: Build Package =========${END}"
+        echo -e "${YELLOW}======== Stage 3: Build Package =========${END}"
 
-    status=0
-    rpmbuild -ba *.spec
-    tmp=$?
-    status=$[tmp + status];
-    checkStatus $status
+        status=0
+        rpmbuild -ba *.spec
+        tmp=$?
+        status=$[tmp + status];
+        checkStatus $status
 
-    echo -e "${YELLOW}======= Stage 4: Regen database =========${END}"
-    createrepo ${RPMPATH}
+        echo -e "${YELLOW}======= Stage 4: Regen database =========${END}"
+        createrepo ${RPMPATH}
+        count2=`find ${RPMPATH}/aarch64/ -type f | wc -l`
+        count2=$[count1 + `find ${RPMPATH}/noarch/ -type f | wc -l`]
+        if [ count2 > count1 ]; then
+            echo "${PWD##*/} OK!" >> ${SRCPATH}/build-log.txt
+            succ=succ+1
+            touch ${SRCPATH}/${pkgname}/success.token
+        else
+            echo "${PWD##*/} Failed!" >> ${SRCPATH}/build-log.txt
+            fail=fail+1
+        fi
+    else
+        skip=skip+1
+    fi
 done < ${PKGLIST}
+echo -e "${GREEN}$succ package(s) built successfully.\n${END}"
+echo -e "${YELLOW}$skip package(s) were skipped.\n${END}"
+echo -e "${RED}$fail package(s) built unsuccessfully.\n${END}"
+echo -e "For detailed info, please check ${SRCPATH}/build-log.txt"
